requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
bcrypt==3.2.2
crypto==1.4.1
cryptography==37.0.2
dnspython==1.15.0
email-validator==1.2.1
firebase-admin==5.2.0
Flask==2.0.3
flask-blueprint==1.3.0
Flask-JWT-Extended==4.4.2
Flask-Limiter==2.5.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
flask-mailman==0.3.0
Flask-Migrate==3.1.0
Flask-Session==0.4.0
flask-talisman==1.0.0
Flask-WTF==1.0.1
flask-csp==0.10
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-cloud-appengine-logging==1.1.3
google-cloud-audit-log==0.2.3
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-kms==2.12.0
google-cloud-logging==3.2.1
google-cloud-secret-manager==2.12.2
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.3
limits==2.6.3
MarkupSafe==2.1.1
oauth2client==4.1.3
pycryptodome==3.14.1
pycryptodomex==3.15.0
PyJWT==2.4.0
Pyrebase4==4.5.0
python-dotenv==0.20.0
python-jwt==3.3.2
requests==2.28.1
schedule==1.1.0
stripe==3.5.0
Werkzeug==2.1.2
wget==3.2
wrapt==1.14.1
WTForms==2.3.3
yarg==0.1.9
zipp==3.8.0
sentry-sdk==1.9.4
python-json-logger==2.0.4